优化del_static方法

+++ User/models.py
@@ -66,7 +66,8 @@ class UserOrder(models.Model):
 
 
 @receiver(post_delete, sender=WxUser)
-def del_shop_static(sender, instance, **kwargs):
-    image_path = instance.avatar.path
-    if image_path and os.path.exists(image_path):
-        os.remove(image_path)
+def del_static(sender, instance, **kwargs):
+    if instance.avatar:
+        image_path = instance.avatar.path
+        if image_path and os.path.exists(image_path):
+            os.remove(image_path)
