+++ comment/models.py
@@ -42,7 +42,8 @@ class CommentUpvote(models.Model):
 
 
 @receiver(post_delete, sender=CommentImage)
-def del_shop_static(sender, instance, **kwargs):
-    image_path = instance.image.path
-    if image_path and os.path.exists(image_path):
-        os.remove(image_path)
+def del_static(sender, instance, **kwargs):
+    if instance.image:
+        image_path = instance.image.path
+        if image_path and os.path.exists(image_path):
+            os.remove(image_path)
